picture-in-picture/leave-picture-in-picture-disable-attribute.optional.html

<!DOCTYPE html>
<title>Test leavepictureinpicture event</title>
<meta name="timeout" content="long">
<script src="/common/media.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/picture-in-picture-helpers.js"></script>
<body></body>
<script>
promise_test(async () => {
  let pictureInPictureWindow;
  const video = await loadVideo();

  video.addEventListener('enterpictureinpicture', event => {
    pictureInPictureWindow = event.pictureInPictureWindow;
  });

  const leavePipPromise = new Promise(resolve => {
    video.addEventListener('leavepictureinpicture', event => {
      assert_equals(pictureInPictureWindow, event.pictureInPictureWindow);
      assert_equals(event.target, video);
      assert_equals(event.bubbles, true);
      assert_equals(event.cancelable, false);
      assert_equals(event.composed, false);
      assert_equals(document.pictureInPictureElement, null);
      resolve();
    });
  });

  await requestPictureInPictureWithTrustedClick(video);
  video.disablePictureInPicture = true;
  await leavePipPromise;
}, 'leavepictureinpicture event is fired if video.disablePictureInPicture is set to true');
</script>
